Modularize Helper and VMs
